arreglos de usuario y producto

+++ src/main/java/com/arkadium/arkadium/Services/UserService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.arkadium.arkadium.Model.User;
+import com.arkadium.arkadium.Repository.CartRepository;
 import com.arkadium.arkadium.Repository.UserRepository;
 
 import jakarta.transaction.Transactional;
@@ -15,6 +16,9 @@ import jakarta.transaction.Transactional;
 public class UserService {
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private CartRepository cartRepository;
 
     public List<User> findAllUsers() {
         return userRepository.findAll();
@@ -37,6 +41,10 @@ public class UserService {
     }
 
     public void deleteUser(Integer id){
+        if (!userRepository.existsById(id)) {
+            throw new RuntimeException("Usuario no encontrado con ID: " + id);
+        }
+        cartRepository.findByUserId(id).ifPresent(cartRepository::delete);
         userRepository.deleteById(id);
     }
 }
